+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Uses the Python limited ABI now and supports python >=3.8.
+- Uses the Python limited ABI now and supports python >=3.11.
 
 ## [0.1.13]